Allow including artifacts from dependencies and reactor projects
If 'includeDependencies' is set to true, the POM dependencies will
be included in the target P2 repository.

If 'includeReactor' is set to true, the project artifacts in the
current reactor will be included in the target P2 repository.

Included projects/dependencies are filtered by type
('includeTypes', defaults jar,bundle) and dependency
scope ('includeScopes', defaults compile,runtime,system).